requirements.txt
aiosmtplib==3.0.2
aiosqlite==0.20.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==5.0.1
Authlib==1.3.2
bcrypt==4.2.1
billiard==4.2.1
cachetools==5.5.0
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==44.0.0
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
environs==11.2.1
exceptiongroup==1.2.2
fastapi==0.115.6
fastapi-jwt==0.3.0
fastapi-jwt-auth==0.5.0
fastapi_cors==0.0.6
google-api-core==2.24.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.66.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jose==1.0.0
kombu==5.4.2
MarkupSafe==3.0.2
marshmallow==3.23.1
oauthlib==3.2.2
packaging==24.2
passlib==1.7.4
prompt_toolkit==3.0.48
proto-plus==1.25.0
protobuf==5.29.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.3
pydantic_core==2.27.1
pydentic==0.0.1.dev3
PyJWT==1.7.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.19
python-stdnum==1.20
redis==5.2.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.17.0
sniffio==1.3.1
sqladmin==0.20.1
SQLAlchemy==2.0.36
starlette==0.41.3
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.32.1
vine==5.1.0
wcwidth==0.2.13
WTForms==3.1.2
stripe~=11.3.0